minio::s3

Module aws_chunked

AWS Chunked encoding with trailing checksum support.

This module implements the aws-chunked content encoding format used by S3 for streaming uploads with trailing checksums. The format allows computing checksums incrementally while streaming data, with the checksum value sent as a trailer at the end of the body.

Unsigned Protocol Format (STREAMING-UNSIGNED-PAYLOAD-TRAILER)

<hex-chunk-size>\r\n
<chunk-data>\r\n
...
0\r\n
x-amz-checksum-<algorithm>:<base64-value>\r\n
\r\n

Signed Protocol Format (STREAMING-AWS4-HMAC-SHA256-PAYLOAD-TRAILER)

<hex-chunk-size>;chunk-signature=<sig>\r\n
<chunk-data>\r\n
...
0;chunk-signature=<final-sig>\r\n
x-amz-checksum-<algorithm>:<base64-value>\r\n
x-amz-trailer-signature:<trailer-sig>\r\n
\r\n

Wire Format vs Canonical Form

Important: There are two different line ending conventions:

Wire format (HTTP protocol): Uses \r\n (CRLF) per RFC 9112 (HTTP/1.1)

Canonical form (for signing): Uses \n (LF) per AWS SigV4 spec

When computing the trailer signature, AWS specifies:

hash('x-amz-checksum-crc32c:sOO8/Q==\n')  // Note: \n not \r\n

But the actual bytes sent over HTTP use CRLF line endings.

Reference: https://docs.aws.amazon.com/AmazonS3/latest/API/sigv4-streaming-trailers.html

Structs

AwsChunkedEncoder

AWS Chunked encoder that wraps data in aws-chunked format with trailing checksum.

RechunkingStream

A stream wrapper that re-chunks incoming data to a fixed chunk size.

SignedAwsChunkedEncoder

AWS Chunked encoder with chunk signing for STREAMING-AWS4-HMAC-SHA256-PAYLOAD-TRAILER.

Functions

calculate_encoded_length

Calculates the encoded length for aws-chunked format.

calculate_signed_encoded_length

Calculates the encoded length for signed aws-chunked format.

default_chunk_size

Returns the default chunk size for aws-chunked encoding.