minio::s3

Module builders

Argument builders for minio::s3::client::Client APIs

Structs

AbortMultipartUpload

Argument for abort_multipart_upload() API

AppendObject

Argument builder for the AppendObject S3 API operation.

AppendObjectContent

Argument builder for the AppendObject S3 API operation.

BucketCommon

Common parameters for bucket operations.

CompleteMultipartUpload

Argument for complete_multipart_upload() API

ComposeObject

Argument builder for CopyObject S3 API operation.

ComposeObjectInternal

ComposeSource

Source object information for compose_object.

ContentStream

CopyObject

Argument builder for CopyObject S3 API operation.

CopyObjectInternal

CopySource

Base argument for object conditional read APIs.

CreateBucket

Argument builder for the CreateBucket S3 API operation.

CreateMultipartUpload

Argument builder for the CreateMultipartUpload S3 API operation.

DeleteObject

Argument builder for the DeleteObject S3 API operation.

DeleteObjectTagging

Argument builder for the DeleteObjectTagging S3 API operation.

DeleteObjects

Argument builder for the DeleteObjects S3 API operation.

DeleteObjectsStreaming

Argument builder for streaming multiple object deletions using the DeleteObjects S3 API operation.

GetBucketLifecycle

Argument builder for the GetBucketLifecycle S3 API operation.

GetBucketTagging

Argument builder for the GetBucketTagging S3 API operation.

GetObject

Argument builder for the GetObject S3 API operation.

GetObjectLegalHold

Argument builder for the GetObjectLegalHold S3 API operation.

GetObjectPrompt

Argument builder for the GetObjectPrompt operation (MinIO extension).

GetObjectRetention

Argument builder for the GetObjectRetention S3 API operation.

GetObjectTagging

Argument builder for the GetObjectTagging S3 API operation.

GetPresignedObjectUrl

This struct constructs the parameters required for the Client::get_presigned_object_url method.

GetPresignedPolicyFormData

Argument builder for generating presigned POST policy for the POST Object S3 API operation.

GetRegion

Argument builder for the GetRegion S3 API operation.

ListBuckets

Argument builder for the ListBuckets S3 API operation.

ListObjects

Argument builder for list_objects() API.

ListenBucketNotification

Argument builder for the ListenBucketNotification

ObjectContent

Object content that can be uploaded or downloaded.

ObjectToDelete

Specifies an object to be deleted.

ObjectsStream

PostPolicy

Post policy information for presigned POST policy form-data.

PutBucketEncryption

Argument builder for the PutBucketEncryption S3 API operation.

PutBucketLifecycle

Argument builder for the PutBucketLifecycle S3 API operation.

PutBucketNotification

Argument builder for the PutBucketNotification S3 API operation.

PutBucketPolicy

Argument builder for the PutBucketPolicy S3 API operation.

PutBucketReplication

Argument builder for the PutBucketReplication S3 API operation.

PutBucketTagging

Argument builder for the PutBucketTagging S3 API operation.

PutBucketVersioning

Argument builder for the PutBucketVersioning S3 API operation.

PutObject

Argument builder for the PutObject S3 API operation.

PutObjectContent

Argument builder for the PutObject S3 API operation with streaming content.

PutObjectLegalHold

Argument builder for the PutObjectLegalHold S3 API operation.

PutObjectLockConfig

Argument builder for the PutObjectLockConfig S3 API operation.

PutObjectRetention

Argument builder for the PutObjectRetention S3 API operation.

PutObjectTagging

Argument builder for the PutObjectTagging S3 API operation.

SelectObjectContent

Argument builder for the SelectObjectContent S3 API operation.

StatObject

Argument builder for the HeadObject S3 API operation.

UploadPart

Argument for upload_part() S3 API

UploadPartCopy

Argument builder for the UploadPartCopy S3 API operation.

Enums

Size

VersioningStatus

Represents the versioning state of an S3 bucket.

Constants

DEFAULT_EXPIRY_SECONDS

The default expiry time in seconds for a GetPresignedObjectUrl.

DEFAULT_PART_SIZE

Default size (in bytes) for a multipart upload part when not explicitly specified.

MAX_DELETE_OBJECTS

Maximum number of objects that can be deleted in a single DeleteObjects request.

MAX_MULTIPART_COUNT

Maximum number of parts allowed in a multipart upload.

MAX_OBJECT_SIZE

Maximum allowed size (in bytes) for a single object upload.

MAX_PART_SIZE

Maximum allowed size (in bytes) for a single multipart upload part.

MIN_PART_SIZE

Minimum allowed size (in bytes) for a multipart upload part (except the last).

Functions

calc_part_info

Returns the size of each part to upload and the total number of parts. The number of parts is None when the object size is unknown.

Type Aliases

AbortMultipartUploadBldr

Builder type for AbortMultipartUpload that is returned by MinioClient::abort_multipart_upload.

AppendObjectBldr

Builder type for AppendObject that is returned by MinioClient::append_object.

AppendObjectContentBldr

Builder type for AppendObjectContent that is returned by MinioClient::append_object_content.

BucketExists

Constructs the parameters for the Client::bucket_exists method.

BucketExistsBldr

Builder type for BucketExists that is returned by MinioClient::bucket_exists.

CompleteMultipartUploadBldr

Builder type for CompleteMultipartUpload that is returned by MinioClient::complete_multipart_upload.

ComposeObjectBldr

Builder type for ComposeObject that is returned by MinioClient::compose_object.

ComposeObjectInternalBldr

Builder type for ComposeObjectInternal that is returned by compose_object_internal method.

CopyObjectBldr

Builder type for CopyObject that is returned by MinioClient::copy_object.

CopyObjectInternalBldr

Builder type for CopyObjectInternal that is returned by copy_object_internal method.

CreateBucketBldr

Builder type for CreateBucket that is returned by MinioClient::create_bucket.

CreateMultipartUploadBldr

Builder type for CreateMultipartUpload that is returned by MinioClient::create_multipart_upload.

DeleteBucket

Argument builder for the DeleteBucket S3 API operation.

DeleteBucketBldr

Builder type for DeleteBucket that is returned by MinioClient::delete_bucket.

DeleteBucketEncryption

Argument builder for the DeleteBucketEncryption S3 API operation.

DeleteBucketEncryptionBldr

Builder type for DeleteBucketEncryption that is returned by MinioClient::delete_bucket_encryption.

DeleteBucketLifecycle

Argument builder for the DeleteBucketLifecycle S3 API operation.

DeleteBucketLifecycleBldr

Builder type for DeleteBucketLifecycle that is returned by MinioClient::delete_bucket_lifecycle.

DeleteBucketNotification

Argument builder for the DeleteBucketNotification S3 API operation. Note: on Amazon S3, a bucket notification is deleted by setting its configuration to empty.

DeleteBucketNotificationBldr

Builder type for DeleteBucketNotification that is returned by MinioClient::delete_bucket_notification.

DeleteBucketPolicy

Argument builder for the DeleteBucketPolicy S3 API operation.

DeleteBucketPolicyBldr

Builder type for DeleteBucketPolicy that is returned by MinioClient::delete_bucket_policy.

DeleteBucketReplication

Argument builder for the DeleteBucketReplication S3 API operation.

DeleteBucketReplicationBldr

Builder type for DeleteBucketReplication that is returned by MinioClient::delete_bucket_replication.

DeleteBucketTagging

Argument builder for the DeleteBucketTagging S3 API operation.

DeleteBucketTaggingBldr

Builder type for DeleteBucketTagging that is returned by MinioClient::delete_bucket_tagging.

DeleteObjectBldr

Builder type for DeleteObject that is returned by MinioClient::delete_object.

DeleteObjectLockConfig

This struct constructs the parameters required for the Client::delete_object_lock_config method.

DeleteObjectLockConfigBldr

Builder type for DeleteObjectLockConfig that is returned by MinioClient::delete_object_lock_config.

DeleteObjectTaggingBldr

Builder type for DeleteObjectTagging that is returned by MinioClient::delete_object_tagging.

DeleteObjectsBldr

Builder type for DeleteObjects that is returned by MinioClient::delete_objects.

GetBucketEncryption

Argument builder for the GetBucketEncryption S3 API operation.

GetBucketEncryptionBldr

Builder type for GetBucketEncryption that is returned by MinioClient::get_bucket_encryption.

GetBucketLifecycleBldr

Builder type for GetBucketLifecycle that is returned by MinioClient::get_bucket_lifecycle.

GetBucketNotification

Argument builder for the GetBucketNotification S3 API operation.

GetBucketNotificationBldr

Builder type for GetBucketNotification that is returned by MinioClient::get_bucket_notification.

GetBucketPolicy

Argument builder for the GetBucketPolicy S3 API operation.

GetBucketPolicyBldr

Builder type for GetBucketPolicy that is returned by MinioClient::get_bucket_policy.

GetBucketReplication

Argument builder for the GetBucketReplication S3 API operation.

GetBucketReplicationBldr

Builder type for GetBucketReplication that is returned by MinioClient::get_bucket_replication.

GetBucketTaggingBldr

Builder type for GetBucketTagging that is returned by MinioClient::get_bucket_tagging.

GetBucketVersioning

Argument builder for the GetBucketVersioning S3 API operation.

GetBucketVersioningBldr

Builder type for GetBucketVersioning that is returned by MinioClient::get_bucket_versioning.

GetObjectBldr

Builder type alias for GetObject.

GetObjectLegalHoldBldr

GetObjectLockConfig

Argument builder for the GetObjectLockConfig S3 API operation.

GetObjectLockConfigBldr

GetObjectPromptBldr

GetObjectRetentionBldr

Builder type alias for GetObjectRetention.

GetObjectTaggingBldr

GetPresignedObjectUrlBldr

Builder type alias for GetPresignedObjectUrl.

GetPresignedPolicyFormDataBldr

Builder type alias for GetPresignedPolicyFormData.

GetRegionBldr

Builder type alias for GetRegion.

ListBucketsBldr

Builder type alias for ListBuckets.

ListObjectBldr

Builder type alias for ListObjects.

ListenBucketNotificationBldr

Builder type alias for ListenBucketNotification.

PutBucketEncryptionBldr

Builder type alias for PutBucketEncryption.

PutBucketLifecycleBldr

Builder type alias for PutBucketLifecycle.

PutBucketNotificationBldr

Builder type alias for PutBucketNotification.

PutBucketPolicyBldr

Builder type for PutBucketPolicy that is returned by MinioClient::put_bucket_policy.

PutBucketReplicationBldr

Builder type for PutBucketReplication that is returned by MinioClient::put_bucket_replication.

PutBucketTaggingBldr

Builder type for PutBucketTagging that is returned by MinioClient::put_bucket_tagging.

PutBucketVersioningBldr

Builder type for PutBucketVersioning that is returned by MinioClient::put_bucket_versioning.

PutObjectBldr

Builder type for PutObject that is returned by MinioClient::put_object.

PutObjectContentBldr

Builder type for PutObjectContent that is returned by MinioClient::put_object_content.

PutObjectLegalHoldBldr

Builder type for PutObjectLegalHold that is returned by MinioClient::put_object_legal_hold.

PutObjectLockConfigBldr

Builder type for PutObjectLockConfig that is returned by MinioClient::put_object_lock_config.

PutObjectRetentionBldr

Builder type for PutObjectRetention that is returned by MinioClient::put_object_retention.

PutObjectTaggingBldr

Builder type for PutObjectTagging that is returned by MinioClient::put_object_tagging.

SelectObjectContentBldr

Builder type for SelectObjectContent that is returned by MinioClient::select_object_content.

StatObjectBldr

Builder type for StatObject that is returned by MinioClient::stat_object.

UploadPartBldr

Builder type for UploadPart that is returned by MinioClient::upload_part.

UploadPartCopyBldr

Builder type for UploadPartCopy that is returned by MinioClient::upload_part_copy.