minio::s3::builders

Struct ObjectContent

pub struct ObjectContent(/* private fields */);

Object content that can be uploaded or downloaded.

Can be constructed from a stream of Bytes, a file path, or a Bytes object.

Implementations

impl ObjectContent

pub fn new_from_stream( r: impl Stream<Item = Result<Bytes, Error>> + Send + 'static, size: impl Into<Size>, ) -> Self

Create a new ObjectContent from a stream of Bytes.

pub async fn to_stream( self, ) -> Result<(Pin<Box<dyn Stream<Item = Result<Bytes, Error>> + Send>>, Size), Error>

pub async fn to_segmented_bytes(self) -> Result<SegmentedBytes, Error>

Load the content into memory and return a SegmentedBytes object.

pub async fn to_file(self, file_path: &Path) -> Result<u64, Error>

Write the content to a file. This function will return the total number of bytes written to the file. It first writes the content to a temporary file and then renames the temporary file to the final file path. The temporary file will be located in the same directory as the final file path.

If the file already exists, it will be replaced. If the parent directory does not exist, an attempt to create it will be made.