minio::s3::builders

Struct PostPolicy

pub struct PostPolicy {
    pub region: Option<Region>,
    pub bucket: BucketName,
    /* private fields */
}

Post policy information for presigned POST policy form-data.

See Post Policy Conditions for condition elements and their usage.

Fields
region: Option<Region>bucket: BucketName
Implementations

impl PostPolicy

pub fn new( bucket: BucketName, expiration: UtcTime, ) -> Result<Self, ValidationErr>

Returns a post policy with given bucket name and expiration.

Examples

use minio::s3::utils::*;
use minio::s3::types::BucketName;
use chrono::Duration;
use minio::s3::builders::PostPolicy;
let expiration = utc_now() + Duration::days(7);
let bucket = BucketName::new("bucket-name").unwrap();
let policy = PostPolicy::new(bucket, expiration).unwrap();

pub fn add_equals_condition( &mut self, element: &str, value: &str, ) -> Result<(), ValidationErr>

Adds equals condition for given element and value

Examples

use minio::s3::utils::*;
use minio::s3::types::BucketName;
use chrono::Duration;
use minio::s3::builders::PostPolicy;
let expiration = utc_now() + Duration::days(7);
let bucket = BucketName::new("my-bucket").unwrap();
let mut policy = PostPolicy::new(bucket, expiration).unwrap();

// Add condition that 'key' (object name) equals to 'bucket-name'
policy.add_equals_condition("key", "bucket-name").unwrap();

pub fn remove_equals_condition(&mut self, element: &str)

Removes equals condition for given element

Examples

use minio::s3::utils::*;
use minio::s3::types::BucketName;
use chrono::Duration;
use minio::s3::builders::PostPolicy;
let expiration = utc_now() + Duration::days(7);
let bucket = BucketName::new("bucket-name").unwrap();
let mut policy = PostPolicy::new(bucket, expiration).unwrap();
policy.add_equals_condition("key", "bucket-name");

policy.remove_equals_condition("key");

pub fn add_starts_with_condition( &mut self, element: &str, value: &str, ) -> Result<(), ValidationErr>

Adds starts-with condition for given element and value

Examples

use minio::s3::utils::*;
use minio::s3::types::BucketName;
use chrono::Duration;
use minio::s3::builders::PostPolicy;
let expiration = utc_now() + Duration::days(7);
let bucket = BucketName::new("bucket-name").unwrap();
let mut policy = PostPolicy::new(bucket, expiration).unwrap();

// Add condition that 'Content-Type' starts with 'image/'
policy.add_starts_with_condition("Content-Type", "image/").unwrap();

pub fn remove_starts_with_condition(&mut self, element: &str)

Removes starts-with condition for given element

Examples

use minio::s3::utils::*;
use minio::s3::types::BucketName;
use chrono::Duration;
use minio::s3::builders::PostPolicy;
let expiration = utc_now() + Duration::days(7);
let bucket = BucketName::new("bucket-name").unwrap();
let mut policy = PostPolicy::new(bucket, expiration).unwrap();
policy.add_starts_with_condition("Content-Type", "image/").unwrap();

policy.remove_starts_with_condition("Content-Type");

pub fn add_content_length_range_condition( &mut self, lower_limit: usize, upper_limit: usize, ) -> Result<(), ValidationErr>

Adds content-length range condition with given lower and upper limits

Examples

use minio::s3::utils::*;
use minio::s3::types::BucketName;
use chrono::Duration;
use minio::s3::builders::PostPolicy;

let expiration = utc_now() + Duration::days(7);
let bucket = BucketName::new("my-bucket").unwrap();
let mut policy = PostPolicy::new(bucket, expiration).unwrap();

// Add condition that 'content-length-range' is between 64kiB to 10MiB
policy.add_content_length_range_condition(64 * 1024, 10 * 1024 * 1024).unwrap();

pub fn remove_content_length_range_condition(&mut self)

Removes content-length range condition