minio::s3

Module client

S3 client to perform bucket and object operations.

HTTP Version Support

The client supports both HTTP/1.1 and HTTP/2. When connecting over TLS, the client will negotiate HTTP/2 via ALPN if the server supports it, otherwise it falls back to HTTP/1.1 gracefully. HTTP/2 provides better throughput for parallel requests through multiplexing.

HTTP/2 support is enabled by default via the http2 feature flag. For HTTP/1.1-only legacy S3-compatible services, you can disable it:

[dependencies]
minio = { version = "0.3", default-features = false, features = ["default-tls", "default-crypto"] }

Re-exports

pub use crate::s3::client::hooks::RequestHooks;

Modules

hooks

Request hooks for intercepting and modifying S3 API requests.

Structs

ConnectionPoolConfig

Configuration for the HTTP connection pool.

Method

The Request Method (VERB)

MinioClient

Simple Storage Service (aka S3) client to perform bucket and object operations.

MinioClientBuilder

Client Builder manufactures a Client using given parameters. Creates a builder given a base URL for the MinIO service or other AWS S3 compatible object storage service.

Response

A Response to a submitted Request.

Statics

DEFAULT_REGION

The default AWS region to be used if no other region is specified.