minio::s3

Module response

Structs

AppendObjectResponse

Response from the append_object API.

BucketExistsResponse

Response from the bucket_exists() API.

CreateBucketResponse

Response from the create_bucket() API.

DeleteBucketEncryptionResponse

Represents the response of the delete_bucket_encryption() API call. This struct contains metadata and information about the bucket whose encryption configuration was removed.

DeleteBucketLifecycleResponse

Represents the response of the delete_bucket_lifecycle() API call. This struct contains metadata and information about the bucket whose lifecycle configuration was removed.

DeleteBucketNotificationResponse

Response from the delete_bucket_notification() API.

DeleteBucketPolicyResponse

Represents the response of the delete_bucket_policy() API call. This struct contains metadata and information about the bucket whose policy was removed.

DeleteBucketReplicationResponse

Response from the delete_bucket_replication() API.

DeleteBucketResponse

Response from the delete_bucket() API.

DeleteBucketTaggingResponse

Represents the response of the delete_bucket_tagging() API call. This struct contains metadata and information about the bucket whose tags were removed.

DeleteError

Error info returned by the S3 API when an object could not be deleted.

DeleteObjectLockConfigResponse

Response from the delete_object_lock_config API call, indicating that the Object Lock configuration has been successfully removed from the specified S3 bucket.

DeleteObjectResponse

DeleteObjectTaggingResponse

Response from the delete_object_tagging API call, indicating that all tags have been successfully removed from a specific object (or object version) in an S3 bucket.

DeleteObjectsResponse

Response of the delete_objects() S3 API.

DeletedObject

Information about an object that was deleted.

GetBucketEncryptionResponse

Response from the get_bucket_encryption API call, providing the default server-side encryption configuration of an S3 bucket.

GetBucketLifecycleResponse

Response from the get_bucket_lifecycle API call, providing the lifecycle configuration of an S3 bucket.

GetBucketNotificationResponse

Response from the get_bucket_notification API call, providing the notification configuration of an S3 bucket.

GetBucketPolicyResponse

Response from the get_bucket_policy API call, providing the bucket policy associated with an S3 bucket.

GetBucketReplicationResponse

Response from the get_bucket_replication API call, providing the replication configuration of an S3 bucket.

GetBucketTaggingResponse

Response from the get_bucket_tagging API call, providing the set of tags associated with an S3 bucket.

GetBucketVersioningResponse

Response from the get_bucket_versioning API call, providing the versioning configuration of a bucket.

GetObjectLegalHoldResponse

Response of get_object_legal_hold() API

GetObjectLockConfigResponse

Response from the get_object_lock_config API call, which retrieves the Object Lock configuration of a bucket.

GetObjectPromptResponse

GetObjectResponse

GetObjectRetentionResponse

Response of get_object_retention() API

GetObjectTaggingResponse

Response of get_object_tags() API

GetPresignedObjectUrlResponse

Response of get_presigned_object_url() API

GetRegionResponse

Response of get_region() API

ListBucketsResponse

Response of list_buckets() API

ListObjectsResponse

Response of list_objects() API

ListenBucketNotificationResponse

Response of listen_bucket_notification() API

PutBucketEncryptionResponse

Response of put_bucket_encryption() API

PutBucketLifecycleResponse

Response of put_bucket_lifecycle() API

PutBucketNotificationResponse

Response of put_bucket_notification() API

PutBucketPolicyResponse

Response of put_bucket_policy() API

PutBucketReplicationResponse

Response of put_bucket_replication() API

PutBucketTaggingResponse

Response of put_bucket_tagging() API

PutBucketVersioningResponse

Response of put_bucket_versioning() API

PutObjectLegalHoldResponse

Response from the put_object_legal_hold API call, indicating that a legal hold has been successfully removed from a specific object version in an S3 bucket.

PutObjectLockConfigResponse

Response of put_object_lock_config() API

PutObjectRetentionResponse

Response of put_object_retention() API

PutObjectTaggingResponse

Response of put_object_tagging() API

S3Response2

Base response struct that contains common functionality for S3 operations

SelectObjectContentResponse

Response of select_object_content() API Response of select_object_content() API

StatObjectResponse

Response from the stat_object API.

Enums

DeleteResult

Result of deleting an object.

Type Aliases

AbortMultipartUploadResponse

Response of abort_multipart_upload() API

CompleteMultipartUploadResponse

Response of complete_multipart_upload() API

ComposeObjectResponse

Represents the response of the compose_object() API call. This struct contains metadata and information about the composed object.

CopyObjectInternalResponse

Internal response type for copy operations

CopyObjectResponse

Represents the response of the copy_object() API call. This struct contains metadata and information about the object being copied.

CreateMultipartUploadResponse

Response of create_multipart_upload() API

PutObjectContentResponse

Response for put_object operations that include object size information

PutObjectResponse

Response of put_object_api() API

UploadPartCopyResponse

Represents the response of the upload_part_copy API call. This struct contains metadata and information about the part being copied during a multipart upload.

UploadPartResponse

Response of upload_part() API