minio::s3::response_traits

Trait HasIsDeleteMarker

pub trait HasIsDeleteMarker: HasS3Fields {
    // Provided method
    fn is_delete_marker(&self) -> Result<bool, ValidationErr> { ... }
}

Provides access to the x-amz-delete-marker header value.

Indicates whether the specified object version that was permanently deleted was (true) or was not (false) a delete marker before deletion. In a simple DELETE, this header indicates whether (true) or not (false) the current version of the object is a delete marker.

Provided Methods

fn is_delete_marker(&self) -> Result<bool, ValidationErr>

Returns true if the object is a delete marker, false otherwise.