minio::s3::response_traits

Trait HasObjectSize

pub trait HasObjectSize: HasS3Fields {
    // Provided method
    fn object_size(&self) -> u64 { ... }
}

Returns the size of the object in bytes, as specified by the x-amz-object-size header.

Provided Methods

fn object_size(&self) -> u64

Returns the size of the object in bytes, as specified by the x-amz-object-size header.