minio::s3::types::traits

Trait ToS3Request

pub trait ToS3Request: Sized {
    // Required method
    fn to_s3request(self) -> Result<S3Request, ValidationErr>;
}

Trait for converting a request builder into a concrete S3 HTTP request.

This trait is implemented by all S3 request builders and serves as an intermediate step in the request execution pipeline. It enables the conversion from a strongly typed request builder into a generic S3Request that can be executed over HTTP.

The S3Api::send method uses this trait to convert request builders into executable HTTP requests before sending them to the S3-compatible service.

See Also

S3Api - The trait that uses ToS3Request as part of its request execution pipeline

FromS3Response - The counterpart trait for converting HTTP responses into typed responses

Required Methods

fn to_s3request(self) -> Result<S3Request, ValidationErr>

Consumes this request builder and returns a S3Request.

This method transforms the request builder into a concrete HTTP request that can be executed against an S3-compatible service. The transformation includes:

Setting the appropriate HTTP method (GET, PUT, POST, etc.)

Building the request URL with path and query parameters

Adding required headers (authentication, content-type, etc.)

Attaching the request body, if applicable

Returns

Result<S3Request, ValidationErr> - The executable S3 request on success, or an error if the request cannot be built correctly.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.