minio::s3

Module utils

Modules

aws_date_format

xml

Enums

ChecksumAlgorithm

S3 checksum algorithms supported by the API

Constants

EMPTY_SHA256

SHA256 hash of empty data.

Functions

b64_encode

Encodes data using base64 algorithm.

check_bucket_name

Validates given bucket name.

check_object_name

Validates given object name.

check_sse

Validates SSE (Server-Side Encryption) settings.

check_ssec

Validates SSE-C (Server-Side Encryption with Customer-Provided Keys) settings.

check_ssec_with_log

Validates SSE-C (Server-Side Encryption with Customer-Provided Keys) settings and logs an error.

compute_checksum

Computes checksum based on the specified algorithm for contiguous byte slices.

compute_checksum_sb

Computes checksum for SegmentedBytes without copying data (zero-copy streaming).

copy_slice

Copies source byte slice into destination byte slice.

crc32

Computes CRC32 of given data using hardware-accelerated SIMD implementation.

crc32_checksum

Computes CRC32 checksum and returns base64-encoded value

crc32c

Computes CRC32C checksum (Castagnoli polynomial) and returns base64-encoded value.

crc64nvme

Computes CRC64-NVME of given data using hardware-accelerated SIMD implementation.

crc64nvme_checksum

Computes CRC64-NVME checksum and returns base64-encoded value

encode_tags

Encodes tags as URL-encoded query parameters for the x-amz-tagging header.

from_http_header_value

Parses HTTP header value to time.

from_iso8601utc

Parses ISO8601 UTC formatted value to time.

get_text_default

Gets default text value of given XML element for given tag.

get_text_option

Gets optional text value of given XML element for given tag.

get_text_result

Gets text value of given XML element for given tag.

hex_encode

Hex-encode a byte slice into a lowercase ASCII string.

insert

Returns the consumed data and inserts a key into it with an empty value.

match_hostname

Checks if given hostname is valid or not.

match_region

Checks if given region is valid or not.

md5sum_hash

Gets base64-encoded MD5 hash of given data.

parse_bool

parse_tags

sha1_hash

Computes SHA1 hash and returns base64-encoded value

sha256_checksum

Computes SHA256 hash and returns base64-encoded value (for checksums, not authentication)

sha256_hash

Gets hex-encoded SHA256 hash of given data.

sha256_hash_sb

to_amz_date

Gets AMZ date value of given time.

to_http_header_value

Gets HTTP header value of given time.

to_iso8601utc

Gets ISO8601 UTC formatted value of given time.

to_signer_date

Gets signer date value of given time.

trim_quotes

Trims leading and trailing quotes from a string.

uint32

Converts data array into 32-bit BigEndian unsigned int.

url_decode

Decodes a URL-encoded string in the application/x-www-form-urlencoded syntax into a string. Note that “+” is decoded to a space character, and “%2B” is decoded to a plus sign.

url_encode

Encodes a string using URL encoding. Note that a whitespace is encoded as “%20” and plus sign is encoded as “%2B”.

urlencode_object_key

utc_now

Gets current UTC time.

Type Aliases

UtcTime

Date and time with UTC timezone.